+++ tests/ByteDev.Http.UnitTests/HttpContentExtensionsTests.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Net.Http;
-using System.Text.Json;
-using System.Threading.Tasks;
+using ByteDev.Http.Content;
 using NUnit.Framework;
 
 namespace ByteDev.Http.UnitTests
@@ -9,50 +7,127 @@ namespace ByteDev.Http.UnitTests
     [TestFixture]
     public class HttpContentExtensionsTests
     {
+        private const string Json = "{ \"IsSuccess\": true } ";
+        private const string Xml = "<data></data>";
+        private const string FormUrlEncoded = "name1=value1&name2=value2";
+
         [TestFixture]
-        public class ReadAsJsonAsync : HttpContentExtensionsTests
+        public class IsJson : HttpContentExtensionsTests
         {
             [Test]
-            public void WhenContentIsNull_ThenThrowException()
+            public void WhenIsNull_ThenThrowException()
             {
-                Assert.ThrowsAsync<ArgumentNullException>(() => HttpContentExtensions.ReadAsJsonAsync<TestDummy>(null));
+                Assert.Throws<ArgumentNullException>(() => HttpContentExtensions.IsJson(null));
             }
 
             [Test]
-            public async Task WhenContentIsJson_ThenReturnObj()
+            public void WhenHasNoContentType_ThenReturnFalse()
             {
-                var dummy = new TestDummy {Name = "John"};
-                var json = JsonSerializer.Serialize(dummy);
+                var sut = new EmptyContent();
 
-                var sut = new StringContent(json);
+                var result = sut.IsJson();
 
-                var result = await sut.ReadAsJsonAsync<TestDummy>();
+                Assert.That(result, Is.False);
+            }
 
-                Assert.That(result.Name, Is.EqualTo(dummy.Name));
+            [Test]
+            public void WhenNonJsonContentType_ThenReturnFalse()
+            {
+                var sut = new XmlContent(Xml);
+
+                var result = sut.IsJson();
+
+                Assert.That(result, Is.False);
+            }
+
+            [Test]
+            public void WhenIsJson_ThenReturnTrue()
+            {
+                var sut = new JsonContent(Json);
+
+                var result = sut.IsJson();
+
+                Assert.That(result, Is.True);
             }
         }
 
         [TestFixture]
-        public class ReadAsXmlAsync : HttpContentExtensionsTests
+        public class IsXml : HttpContentExtensionsTests
         {
             [Test]
-            public void WhenContentIsNull_ThenThrowException()
+            public void WhenIsNull_ThenThrowException()
             {
-                Assert.ThrowsAsync<ArgumentNullException>(() => HttpContentExtensions.ReadAsJsonAsync<TestDummy>(null));
+                Assert.Throws<ArgumentNullException>(() => HttpContentExtensions.IsXml(null));
             }
 
             [Test]
-            public async Task WhenContentIsXml_ThenReturnObj()
+            public void WhenHasNoContentType_ThenReturnFalse()
             {
-                var dummy = new TestDummy {Name = "John"};
+                var sut = new EmptyContent();
 
-                var xml = XmlDataSerializer.Serialize(dummy);
+                var result = sut.IsXml();
 
-                var sut = new StringContent(xml);
+                Assert.That(result, Is.False);
+            }
 
-                var result = await sut.ReadAsXmlAsync<TestDummy>();
+            [Test]
+            public void WhenNonXmlContentType_ThenReturnFalse()
+            {
+                var sut = new JsonContent(Json);
 
-                Assert.That(result.Name, Is.EqualTo(dummy.Name));
+                var result = sut.IsXml();
+
+                Assert.That(result, Is.False);
+            }
+
+            [Test]
+            public void WhenIsXml_ThenReturnTrue()
+            {
+                var sut = new XmlContent(Xml);
+
+                var result = sut.IsXml();
+
+                Assert.That(result, Is.True);
+            }
+        }
+
+        [TestFixture]
+        public class IsFormUrlEncoded : HttpContentExtensionsTests
+        {
+            [Test]
+            public void WhenIsNull_ThenThrowException()
+            {
+                Assert.Throws<ArgumentNullException>(() => HttpContentExtensions.IsFormUrlEncoded(null));
+            }
+
+            [Test]
+            public void WhenHasNoContentType_ThenReturnFalse()
+            {
+                var sut = new EmptyContent();
+
+                var result = sut.IsFormUrlEncoded();
+
+                Assert.That(result, Is.False);
+            }
+
+            [Test]
+            public void WhenNonXmlContentType_ThenReturnFalse()
+            {
+                var sut = new JsonContent(Json);
+
+                var result = sut.IsFormUrlEncoded();
+
+                Assert.That(result, Is.False);
+            }
+
+            [Test]
+            public void WhenIsFormUrlEncoded_ThenReturnTrue()
+            {
+                var sut = new FormUrlEncodedContent(FormUrlEncoded);
+
+                var result = sut.IsFormUrlEncoded();
+
+                Assert.That(result, Is.True);
             }
         }
     }